@@ -5,7 +5,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -143,7 +142,7 @@ func (s *ObserverSuite) SetUpSuite(c *C) {
 			_, err := rw.Write([]byte(`{ "jsonrpc": "2.0", "id": "E7FDA9DE4D0AD37D823813CB5BC0D6E69AB0D41BB666B65B965D12D24A3AE83C", "result": { "height": "1", "txhash": "AAAA000000000000000000000000000000000000000000000000000000000000", "logs": [{"success": "true", "log": ""}] } }`))
 			c.Assert(err, IsNil)
 		} else if strings.HasPrefix(req.RequestURI, thorclient.MimirEndpoint) {
-			buf, err := ioutil.ReadFile("../../test/fixtures/endpoints/mimir/mimir.json")
+			buf, err := os.ReadFile("../../test/fixtures/endpoints/mimir/mimir.json")
 			c.Assert(err, IsNil)
 			_, err = rw.Write(buf)
 			c.Assert(err, IsNil)
