@@ -4,11 +4,12 @@ import (
 	_ "embed"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"math/big"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"os"
 	"strings"
 	"time"
 
@@ -128,7 +129,7 @@ func (s *BlockScannerTestSuite) TestNewBlockScanner(c *C) {
 	storage, err := blockscanner.NewBlockScannerStorage("")
 	c.Assert(err, IsNil)
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		body, err := ioutil.ReadAll(req.Body)
+		body, err := io.ReadAll(req.Body)
 		c.Assert(err, IsNil)
 		type RPCRequest struct {
 			JSONRPC string          `json:"jsonrpc"`
@@ -194,7 +195,7 @@ func (s *BlockScannerTestSuite) TestProcessBlock(c *C) {
 		case strings.HasPrefix(req.RequestURI, thorclient.AuthAccountEndpoint):
 			httpTestHandler(c, rw, "../../../../test/fixtures/endpoints/auth/accounts/template.json")
 		default:
-			body, err := ioutil.ReadAll(req.Body)
+			body, err := io.ReadAll(req.Body)
 			c.Assert(err, IsNil)
 			defer func() {
 				c.Assert(req.Body.Close(), IsNil)
@@ -276,7 +277,7 @@ func httpTestHandler(c *C, rw http.ResponseWriter, fixture string) {
 	case "500":
 		rw.WriteHeader(http.StatusInternalServerError)
 	default:
-		content, err = ioutil.ReadFile(fixture)
+		content, err = os.ReadFile(fixture)
 		if err != nil {
 			c.Fatal(err)
 		}
@@ -300,7 +301,7 @@ func (s *BlockScannerTestSuite) TestGetTxInItem(c *C) {
 		case strings.HasPrefix(req.RequestURI, thorclient.NodeAccountEndpoint):
 			httpTestHandler(c, rw, "../../../../test/fixtures/endpoints/nodeaccount/template.json")
 		default:
-			body, err := ioutil.ReadAll(req.Body)
+			body, err := io.ReadAll(req.Body)
 			c.Assert(err, IsNil)
 			type RPCRequest struct {
 				JSONRPC string          `json:"jsonrpc"`
@@ -499,7 +500,7 @@ func (s *BlockScannerTestSuite) TestUpdateGasPrice(c *C) {
 	storage, err := blockscanner.NewBlockScannerStorage("")
 	c.Assert(err, IsNil)
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		body, err := ioutil.ReadAll(req.Body)
+		body, err := io.ReadAll(req.Body)
 		c.Assert(err, IsNil)
 		type RPCRequest struct {
 			JSONRPC string          `json:"jsonrpc"`
