@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -91,7 +91,7 @@ func (s *BitcoinCashSignerSuite) SetUpTest(c *C) {
 			r := struct {
 				Method string `json:"method"`
 			}{}
-			buf, err := ioutil.ReadAll(req.Body)
+			buf, err := io.ReadAll(req.Body)
 			c.Assert(err, IsNil)
 			if len(buf) == 0 {
 				return
