@@ -2,7 +2,6 @@ package dogecoin
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -169,7 +168,7 @@ func (s *DogecoinSuite) TearDownTest(_ *C) {
 }
 
 func httpTestHandler(c *C, rw http.ResponseWriter, fixture string) {
-	content, err := ioutil.ReadFile(fixture)
+	content, err := os.ReadFile(fixture)
 	if err != nil {
 		c.Fatal(err)
 	}
@@ -732,7 +731,7 @@ func (s *DogecoinSuite) TestOnObservedTxIn(c *C) {
 func (s *DogecoinSuite) TestProcessReOrg(c *C) {
 	// can't get previous block meta should not error
 	var result btcjson.GetBlockVerboseTxResult
-	blockContent, err := ioutil.ReadFile("../../../../test/fixtures/doge/block.json")
+	blockContent, err := os.ReadFile("../../../../test/fixtures/doge/block.json")
 	c.Assert(err, IsNil)
 	c.Assert(json.Unmarshal(blockContent, &result), IsNil)
 	// should not trigger re-org process
