@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"os/signal"
 	"strings"
@@ -265,7 +264,7 @@ func getLocalPreParam(file string) *btsskeygen.LocalPreParams {
 		return nil
 	}
 	// #nosec G304 this is to read a file provided by a start up parameter , it will not be any random user input
-	buf, err := ioutil.ReadFile(file)
+	buf, err := os.ReadFile(file)
 	if err != nil {
 		log.Fatal().Msgf("fail to read file:%s", file)
 		return nil
