@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 	"os/exec"
@@ -201,7 +201,7 @@ func (op *OpCheck) Execute(_ *os.Process, logs chan string) error {
 	}
 
 	// read response
-	buf, err := ioutil.ReadAll(resp.Body)
+	buf, err := io.ReadAll(resp.Body)
 	if err != nil {
 		log.Err(err).Msg("failed to read response")
 		return err
